Resolved merge conflicts by keeping main branch versions for core system files.
Integrated Jules improvements while maintaining system stability.

+++ src/network/tcp_udp.rs
@@ -324,7 +324,7 @@ impl TCPConnection for SimpleSocket {
     }
 
     fn get_state(&self) -> TCPState {
-        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst)) }
+        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst) as u32) }
     }
 }
